# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SmartZim Learning Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      parameters:
        - name: grade
          in: query
          required: false
          schema:
            type: string
        - name: teacherId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
    post:
      operationId: createAssignment
      tags: [assignments]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssignmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /assignments/{id}:
    get:
      operationId: getAssignment
      tags: [assignments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
    delete:
      operationId: deleteAssignment
      tags: [assignments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /assignments/{id}/submissions:
    get:
      operationId: listSubmissions
      tags: [assignments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
    post:
      operationId: submitAssignment
      tags: [assignments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAssignmentInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/{id}/grade:
    post:
      operationId: gradeSubmission
      tags: [assignments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeSubmissionInput"
      responses:
        "200":
          description: Graded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    User:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        role: { type: string }
        grade: { type: ["string", "null"] }
        curriculum: { type: string }
        subjects:
          type: array
          items: { type: string }
        school: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        status: { type: string }
        subscriptionStatus: { type: string }
        subscriptionExpiry: { type: ["string", "null"], format: date-time }
        trialStartDate: { type: ["string", "null"], format: date-time }
        referralCode: { type: ["string", "null"] }
        profilePhotoUrl: { type: ["string", "null"] }
        coverPhotoUrl: { type: ["string", "null"] }
        studyStreak: { type: integer }
        totalStudyMinutes: { type: integer }
        rejectionReason: { type: ["string", "null"] }
        isSuperAdmin: { type: boolean }
        createdAt: { type: string, format: date-time }
        lastActiveAt: { type: ["string", "null"], format: date-time }
      required:
        - id
        - name
        - email
        - role
        - status
        - subscriptionStatus
        - studyStreak
        - totalStudyMinutes
        - isSuperAdmin
        - createdAt
...
    Assignment:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        instructions: { type: string }
        subject: { type: string }
        grade: { type: string }
        deadline: { type: string, format: date-time }
        fileUrl: { type: ["string", "null"] }
        teacherId: { type: integer }
        teacherName: { type: ["string", "null"] }
        status: { type: string }
        createdAt: { type: string, format: date-time }
        submissionCount: { type: integer }
        mySubmission:
          oneOf:
            - $ref: "#/components/schemas/Submission"
            - type: "null"
      required:
        - id
        - title
        - instructions
        - subject
        - grade
        - deadline
        - teacherId
        - status
        - createdAt
        - submissionCount
    CreateAssignmentInput:
      type: object
      properties:
        title: { type: string }
        instructions: { type: string }
        subject: { type: string }
        grade: { type: string }
        deadline: { type: string, format: date-time }
        fileUrl: { type: ["string", "null"] }
        status: { type: string }
      required: [title, instructions, subject, grade, deadline]
    Submission:
      type: object
      properties:
        id: { type: integer }
        assignmentId: { type: integer }
        assignmentTitle: { type: ["string", "null"] }
        studentId: { type: integer }
        studentName: { type: string }
        textResponse: { type: ["string", "null"] }
        fileUrl: { type: ["string", "null"] }
        submittedAt: { type: string, format: date-time }
        grade: { type: ["integer", "null"] }
        feedback: { type: ["string", "null"] }
        gradedAt: { type: ["string", "null"], format: date-time }
        gradedBy: { type: ["string", "null"] }
      required: [id, assignmentId, studentId, studentName, submittedAt]
    SubmitAssignmentInput:
      type: object
      properties:
        textResponse: { type: ["string", "null"] }
        fileUrl: { type: ["string", "null"] }
    GradeSubmissionInput:
      type: object
      properties:
        grade: { type: integer }
        feedback: { type: string }
      required: [grade, feedback]
...